paymentsense-coding-challenge-api/Paymentsense.Coding.Challenge.Api/Controllers/CountryController.cs
using System;

using Microsoft.AspNetCore.Mvc;

using Paymentsense.Coding.Challenge.Api.Services;

namespace Paymentsense.Coding.Challenge.Api.Controllers
{
    [ApiController]
    [Route("[controller]")]
    public class CountryController : ControllerBase
    {
        private readonly ICountryService _countryService;

        public CountryController(
            ICountryService countryService)
        {
            _countryService = countryService ?? throw new ArgumentNullException(nameof(countryService));
        }

        [HttpGet]
        public ActionResult<string> Get()
        {
            return Ok("Paymentsense Coding Challenge!");
        }

        [HttpGet]
        [Route("/names")]
        public ActionResult<string[]> GetNames()
        {
            throw new NotImplementedException();
        }
    }
}